first online commit

+++ app/Models/Review.php
@@ -2,10 +2,16 @@
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Database\Eloquent\SoftDeletes;
+use Sofa\Revisionable\Laravel\RevisionableTrait; // trait
+use Sofa\Revisionable\Revisionable; // interface
 use Lang;
 
-class Review extends Model {
-
+class Review extends Model implements Revisionable{
+	
+	use SoftDeletes;
+	use RevisionableTrait;
+	protected $dates = ['deleted_at'];
 	protected $table = 'reviews';
 	/**
 	* Official SLMTA?
@@ -66,7 +72,7 @@ class Review extends Model {
 	 	return $this->belongsToMany('App\Models\User', 'review_assessors', 'review_id', 'assessor_id');
 	}
 	//	Set auditors for the review
-	public function setAssessors($field){
+	public function setAssessors($field, $review_user_id){
 
 		$fieldAdded = array();
 		$reviewId = 0;	
@@ -84,7 +90,7 @@ class Review extends Model {
 
 		}
 		// Delete existing parent-child mappings
-		DB::table('review_assessors')->where('review_id', '=', $reviewId)->delete();
+		DB::table('review_assessors')->where('review_id', '=', $reviewId)->where('assessor_id', '!=', $review_user_id)->delete();
 
 		// Add the new mapping
 		DB::table('review_assessors')->insert($fieldAdded);
